feat(inventory-sync): integer-thousandths quantities for fridge_items/transactions
fam-lem.30.1: move MAX_INVENTORY_QUANTITY_UNITS, isPositiveIntegerThousandths
and isNonNegativeIntegerThousandths from inventory-lifecycle.ts into their
declared owner src/lib/inventory-quantity.ts (contract.md section 3).
Consolidate the duplicated fromInventoryQuantityUnits(toInventoryQuantityUnits(x))
precision round-trip (4 near-identical occurrences across 3 files) into a
single assertInventoryQuantityPrecision helper on the same owner.

fam-lem.30.2 (partial): add mapFridgeItemRow as the single named
persistence/view boundary in use-inventory-items.ts, currently an identity
pass-through until the local SQLite write-path cutover lands.

fam-lem.30.6: switch the Supabase-authoritative quantity columns
(fridge_items.quantity/package_size, transactions.quantity/origin_quantity)
from numeric(10,3) to bigint, and the 7 RPCs that read/write them
(adjust/correct/reverse_inventory_quantity_transaction, move/reverse_move,
split/merge_undo_fridge_item_open) to bigint parameters, dropping the
trunc(...,3)/NaN precision checks now owned exclusively by
inventory-quantity.ts. Local dev data was disposable (app unshipped), so the
generated migration does a direct type cast rather than a value-preserving
scale; a second declarative-sync pass was needed to catch a pg-delta quirk
that leaves anon with stale EXECUTE grants after a DROP+CREATE signature
change (same pattern as the earlier fix_*_anon_revoke_gap migrations).
Updated the 8 affected pgTAP suites to integer-thousandths literals and
replaced 3 RPC-level "reject >3 decimals" tests that the bigint parameter
type makes unrepresentable with tests for the validations the RPCs still
own (zero delta, equal expected/new quantity, open > source quantity).

Verification: 131/131 pgTAP tests green, db:advisors clean, db:diff empty
after reset, db:types regenerated, focused Jest suites green, typecheck and
biome clean on all touched files.

+++ supabase/schemas/20_privileges.sql
@@ -154,20 +154,20 @@ revoke execute on function public.create_household(text) from public, anon;
 grant execute on function public.create_household(text) to authenticated;
 
 revoke execute on function public.adjust_fridge_item_quantity(
-  uuid, uuid, uuid, uuid, numeric, timestamptz
+  uuid, uuid, uuid, uuid, bigint, timestamptz
 ) from public;
 revoke all on function public.adjust_fridge_item_quantity(
-  uuid, uuid, uuid, uuid, numeric, timestamptz
+  uuid, uuid, uuid, uuid, bigint, timestamptz
 ) from anon;
 grant execute on function public.adjust_fridge_item_quantity(
-  uuid, uuid, uuid, uuid, numeric, timestamptz
+  uuid, uuid, uuid, uuid, bigint, timestamptz
 ) to authenticated;
 
 revoke execute on function public.correct_fridge_item_quantity(
-  uuid, uuid, uuid, uuid, numeric, numeric, timestamptz
+  uuid, uuid, uuid, uuid, bigint, bigint, timestamptz
 ) from public, anon;
 grant execute on function public.correct_fridge_item_quantity(
-  uuid, uuid, uuid, uuid, numeric, numeric, timestamptz
+  uuid, uuid, uuid, uuid, bigint, bigint, timestamptz
 ) to authenticated;
 
 revoke execute on function public.reverse_inventory_quantity_transaction(
@@ -178,24 +178,24 @@ grant execute on function public.reverse_inventory_quantity_transaction(
 ) to authenticated;
 
 revoke execute on function public.move_fridge_item(
-  uuid, uuid, uuid, uuid, uuid, numeric, uuid, uuid, timestamptz
+  uuid, uuid, uuid, uuid, uuid, bigint, uuid, uuid, timestamptz
 ) from public, anon;
 grant execute on function public.move_fridge_item(
-  uuid, uuid, uuid, uuid, uuid, numeric, uuid, uuid, timestamptz
+  uuid, uuid, uuid, uuid, uuid, bigint, uuid, uuid, timestamptz
 ) to authenticated;
 
 revoke execute on function public.reverse_move_fridge_item(
-  uuid, uuid, uuid, uuid, uuid, uuid, numeric, uuid, uuid, timestamptz, text
+  uuid, uuid, uuid, uuid, uuid, uuid, bigint, uuid, uuid, timestamptz, text
 ) from public, anon;
 grant execute on function public.reverse_move_fridge_item(
-  uuid, uuid, uuid, uuid, uuid, uuid, numeric, uuid, uuid, timestamptz, text
+  uuid, uuid, uuid, uuid, uuid, uuid, bigint, uuid, uuid, timestamptz, text
 ) to authenticated;
 
 revoke execute on function public.split_fridge_item_open(
-  uuid, uuid, uuid, uuid, numeric, numeric, timestamptz, date, boolean, timestamptz
+  uuid, uuid, uuid, uuid, bigint, bigint, timestamptz, date, boolean, timestamptz
 ) from public, anon;
 grant execute on function public.split_fridge_item_open(
-  uuid, uuid, uuid, uuid, numeric, numeric, timestamptz, date, boolean, timestamptz
+  uuid, uuid, uuid, uuid, bigint, bigint, timestamptz, date, boolean, timestamptz
 ) to authenticated;
 
 revoke execute on function public.merge_undo_fridge_item_open(
